+++ backend/database/migrations/add_new_fields.sql
@@ -2,7 +2,7 @@ USE railway;
 
 ALTER TABLE applications 
 ADD COLUMN job_id VARCHAR(100) DEFAULT NULL AFTER id,
-ADD COLUMN salary INT DEFAULT NULL AFTER applied_date,
+ADD COLUMN salary DECIMAL(6,2) DEFAULT NULL AFTER applied_date,
 ADD COLUMN status_updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP AFTER status;
 
 CREATE INDEX idx_salary ON applications(salary);
